fixed an issue with the zonetree using of the containswhat query misses most of cellcontent so use intersectswithquery insteadItems will mow display proper radial menus when dropped in a house People on the administrator list are able to pick up items in a house Only people on the administator list are able to drop items in a house implemented itemmovement and (buggy) item rotation Fixed an issue were NPC Inventories were created with faulty storage values (for non db based NPCs)

+++ src/Utils/Timer.cpp
@@ -14,7 +14,7 @@ Copyright (c) 2006 - 2010 The swgANH Team
 
 //==============================================================================================================================
 
-Timer::Timer(uint32 id,TimerCallback* callback,uint32 interval, void* container)
+Timer::Timer(uint32 id,TimerCallback* callback,uint64 interval, void* container)
 : mContainer(container)
 , mCallback(callback)
 , mId(id)
@@ -23,7 +23,7 @@ Timer::Timer(uint32 id,TimerCallback* callback,uint32 interval, void* container)
 	if(mInterval < 100)
 		mInterval = 100;
 
-	mLastTick = static_cast<uint32>(Anh_Utils::Clock::getSingleton()->getLocalTime());
+	mLastTick = Anh_Utils::Clock::getSingleton()->getLocalTime();
 
     boost::thread t(&Timer::Run, this);
     mThread = boost::move(t);
@@ -50,7 +50,7 @@ void Timer::Run()
 		if(currentTick - mLastTick >= mInterval)
 		{
 			mCallback->handleTimer(mId,mContainer);
-			mLastTick = static_cast<uint32>(currentTick);
+			mLastTick = currentTick;
 		}
 
         boost::this_thread::sleep(boost::posix_time::milliseconds(10));
